:feat: refined defaults
